Remove error wrapper.
